+++ src/modules/internet/nanoftp.c
@@ -575,8 +575,8 @@ RxmlNanoFTPGetMore(void *ctx) {
     /*
      * Read the amount left on the control connection
      */
-    if ((len = recv(ctxt->controlFd, &ctxt->controlBuf[ctxt->controlBufIndex],
-		    size, 0)) < 0) {
+    if ((len = (int) recv(ctxt->controlFd, &ctxt->controlBuf[ctxt->controlBufIndex],
+			  size, 0)) < 0) {
 	RxmlMessage(1, "recv failed");
 	closesocket(ctxt->controlFd); ctxt->controlFd = -1;
 	ctxt->controlFd = -1;
@@ -638,7 +638,7 @@ get_more:
 
     RxmlMessage(0, "\n<<<\n%s\n--\n", ptr);
     while (ptr < end) {
-	cur = RxmlNanoFTPParseResponse(ptr, end - ptr);
+	cur = RxmlNanoFTPParseResponse(ptr, (int)(end - ptr));
 	if (cur > 0) {
 	    /*
 	     * Successfully scanned the control code, scratch
@@ -648,7 +648,7 @@ get_more:
 	    res = cur;
 	    if(res == 150) RxmlFindLength(ctxt, ptr);
 	    ptr += 3;
-	    ctxt->controlBufAnswer = ptr - ctxt->controlBuf;
+	    ctxt->controlBufAnswer = (int)(ptr - ctxt->controlBuf);
 	    while ((ptr < end) && (*ptr != '\n')) ptr++;
 	    if (*ptr == '\n') ptr++;
 	    if (*ptr == '\r') ptr++;
@@ -663,7 +663,7 @@ get_more:
     }
 
     if (res < 0) goto get_more;
-    ctxt->controlBufIndex = ptr - ctxt->controlBuf;
+    ctxt->controlBufIndex = (int)(ptr - ctxt->controlBuf);
     ptr = &ctxt->controlBuf[ctxt->controlBufIndex];
     RxmlMessage(1, "\n---\n%s\n--\n", ptr);
     RxmlMessage(1, "Got %d", res);
@@ -730,14 +730,14 @@ RxmlNanoFTPSendUser(void *ctx) {
     RxmlNanoFTPCtxtPtr ctxt = (RxmlNanoFTPCtxtPtr) ctx;
     char buf[200];
     int len;
-    int res;
+    ssize_t res;
 
     if (ctxt->user == NULL)
 	snprintf(buf, sizeof(buf), "USER anonymous\r\n");
     else
 	snprintf(buf, sizeof(buf), "USER %s\r\n", ctxt->user);
     buf[sizeof(buf) - 1] = 0;
-    len = strlen(buf);
+    len = (int) strlen(buf);
     RxmlMessage(0, "%s", buf);
     res = send(ctxt->controlFd, buf, len, 0);
     if (res < 0) {
@@ -756,14 +756,14 @@ RxmlNanoFTPSendPasswd(void *ctx) {
     RxmlNanoFTPCtxtPtr ctxt = (RxmlNanoFTPCtxtPtr) ctx;
     char buf[200];
     int len;
-    int res;
+    ssize_t res;
 
     if (ctxt->passwd == NULL)
 	snprintf(buf, sizeof(buf), "PASS anonymous@\r\n");
     else
 	snprintf(buf, sizeof(buf), "PASS %s\r\n", ctxt->passwd);
     buf[sizeof(buf) - 1] = 0;
-    len = strlen(buf);
+    len = (int) strlen(buf);
     RxmlMessage(0, "%s", buf);
     res = send(ctxt->controlFd, buf, len, 0);
     if (res < 0) {
@@ -792,7 +792,7 @@ RxmlNanoFTPQuit(void *ctx) {
     if ((ctxt == NULL) || (ctxt->controlFd < 0)) return(-1);
 
     snprintf(buf, sizeof(buf), "QUIT\r\n");
-    len = strlen(buf);
+    len = (int) strlen(buf);
     RxmlMessage(0, "%s", buf);
     send(ctxt->controlFd, buf, len, 0);
     return(0);
@@ -812,7 +812,7 @@ RxmlNanoFTPConnect(void *ctx) {
     RxmlNanoFTPCtxtPtr ctxt = (RxmlNanoFTPCtxtPtr) ctx;
     struct hostent *hp;
     int port;
-    int res;
+    ssize_t res;
 
     if (ctxt == NULL)
 	return(-1);
@@ -917,7 +917,7 @@ RxmlNanoFTPConnect(void *ctx) {
 	     */
 	    snprintf(buf, sizeof(buf), "USER %s\r\n", proxyUser);
 	    buf[sizeof(buf) - 1] = 0;
-	    len = strlen(buf);
+	    len = (int) strlen(buf);
 	    RxmlMessage(0, "%s", buf);
 	    res = send(ctxt->controlFd, buf, len, 0);
 	    if (res < 0) {
@@ -937,7 +937,7 @@ RxmlNanoFTPConnect(void *ctx) {
 		    else
 			snprintf(buf, sizeof(buf), "PASS anonymous@\r\n");
 		    buf[sizeof(buf) - 1] = 0;
-		    len = strlen(buf);
+		    len = (int) strlen(buf);
 		    RxmlMessage(0, "%s", buf);
 		    res = send(ctxt->controlFd, buf, len, 0);
 		    if (res < 0) {
@@ -976,7 +976,7 @@ RxmlNanoFTPConnect(void *ctx) {
 		/* Using SITE command */
 		snprintf(buf, sizeof(buf), "SITE %s\r\n", ctxt->hostname);
 		buf[sizeof(buf) - 1] = 0;
-		len = strlen(buf);
+		len = (int) strlen(buf);
 		RxmlMessage(0, "%s", buf);
 		res = send(ctxt->controlFd, buf, len, 0);
 		if (res < 0) {
@@ -1005,7 +1005,7 @@ RxmlNanoFTPConnect(void *ctx) {
 		    snprintf(buf, sizeof(buf), "USER %s@%s\r\n",
 				   ctxt->user, ctxt->hostname);
 		buf[sizeof(buf) - 1] = 0;
-		len = strlen(buf);
+		len = (int) strlen(buf);
 		RxmlMessage(0, "%s", buf);
 		res = send(ctxt->controlFd, buf, len, 0);
 		if (res < 0) {
@@ -1026,7 +1026,7 @@ RxmlNanoFTPConnect(void *ctx) {
 
 		    snprintf(buf, sizeof(buf), "PASS %s\r\n", ctxt->passwd);
 		buf[sizeof(buf) - 1] = 0;
-		len = strlen(buf);
+		len = (int) strlen(buf);
 		RxmlMessage(0, "%s", buf);
 		res = send(ctxt->controlFd, buf, len, 0);
 		if (res < 0) {
@@ -1140,7 +1140,7 @@ RxmlNanoFTPGetConnection(void *ctx) {
 
     if (ctxt->passive) {
 	snprintf (buf, sizeof(buf), "PASV\r\n");
-	len = strlen(buf);
+	len = (int) strlen(buf);
 #ifdef DEBUG_FTP
 	RxmlMessage(0, "%s", buf);
 #endif
@@ -1203,7 +1203,7 @@ RxmlNanoFTPGetConnection(void *ctx) {
 	       adp[0] & 0xff, adp[1] & 0xff, adp[2] & 0xff, adp[3] & 0xff,
 	       portp[0] & 0xff, portp[1] & 0xff);
 	buf[sizeof(buf) - 1] = 0;
-	len = strlen(buf);
+	len = (int) strlen(buf);
 #ifdef DEBUG_FTP
 	RxmlMessage(1, "%s", buf);
 #endif
@@ -1251,7 +1251,7 @@ RxmlNanoFTPGetSocket(void *ctx, const char *filename) {
 	return(-1);
 
     snprintf(buf, sizeof(buf), "TYPE I\r\n");
-    len = strlen(buf);
+    len = (int) strlen(buf);
 #ifdef DEBUG_FTP
     RxmlMessage(0, "%s", buf);
 #endif
@@ -1271,7 +1271,7 @@ RxmlNanoFTPGetSocket(void *ctx, const char *filename) {
     else
 	snprintf(buf, sizeof(buf), "RETR %s\r\n", filename);
     buf[sizeof(buf) - 1] = 0;
-    len = strlen(buf);
+    len = (int) strlen(buf);
 #ifdef DEBUG_FTP
     RxmlMessage(0, "%s", buf);
 #endif
